feat: refactor main archive to classes to create scenes. Add start screen.
